Move do-transform.js
